requirements.txt
flask
python-swiftclient
python-keystoneclient
pydes